chore: rename python namespace to match the project name
