+++ tests/python/test_docker.py
@@ -1,4 +1,4 @@
-"""Test the houdini_docker_builder.docker module."""
+"""Test the hython_docker_image_builder.docker module."""
 
 # Standard Library
 import subprocess
@@ -6,20 +6,20 @@ import subprocess
 # Third Party
 import pytest
 
-# houdini_docker_builder
-from houdini_docker_builder import docker
+# hython_docker_image_builder
+from hython_docker_image_builder import docker
 
 
 def test_build_full_tag_name():
-    """Test houdini_docker_builder.docker.build_full_tag_name()."""
+    """Test hython_docker_image_builder.docker.build_full_tag_name()."""
     result = docker.build_full_tag_name("name/repo", "20.0")
     assert result == "name/repo:20.0"
 
 
 @pytest.mark.parametrize("exists", (False, True))
 def test_check_tag_exists(mocker, fp, exists):
-    """Test houdini_docker_builder.docker.check_tag_exists()."""
-    mock_build_name = mocker.patch("houdini_docker_builder.docker.build_full_tag_name")
+    """Test hython_docker_image_builder.docker.check_tag_exists()."""
+    mock_build_name = mocker.patch("hython_docker_image_builder.docker.build_full_tag_name")
 
     def callback_function(*args, **kwargs):
         if not exists:
