+++ tox.ini
@@ -31,34 +31,34 @@ commands =
 
 [testenv:ruff-check]
 deps = ruff
-commands = ruff check --preview bin/ python/houdini_docker_builder/ tests/
+commands = ruff check --preview bin/ python/hython_docker_image_builder/ tests/
 
 [testenv:ruff-format-check]
 deps = {[testenv:ruff-check]deps}
-commands = ruff format --preview --check bin/ python/houdini_docker_builder/ tests/
+commands = ruff format --preview --check bin/ python/hython_docker_image_builder/ tests/
 
 [testenv:ruff-format-fix]
 deps = {[testenv:ruff-check]deps}
-commands = ruff format --preview bin/ python/houdini_docker_builder/ tests/
+commands = ruff format --preview bin/ python/hython_docker_image_builder/ tests/
 
 [testenv:isort-check]
 deps = isort
-commands = isort --check bin/ python/houdini_docker_builder/ tests/
+commands = isort --check bin/ python/hython_docker_image_builder/ tests/
 
 [testenv:isort-run]
 deps = {[testenv:isort-check]deps}
-commands = isort bin/ python/houdini_docker_builder tests/
+commands = isort bin/ python/hython_docker_image_builder tests/
 
 [testenv:mypy]
 deps =
     mypy[toml]
     types-requests
-commands = mypy bin python/houdini_docker_builder
+commands = mypy bin python/hython_docker_image_builder
 
 [testenv:docstring-check]
 deps = pydocstringformatter
-commands = pydocstringformatter --exit-code bin/ python/houdini_docker_builder/
+commands = pydocstringformatter --exit-code bin/ python/hython_docker_image_builder/
 
 [testenv:docstring-run]
 deps = {[testenv:docstring-check]deps}
-commands = pydocstringformatter --write bin/ python/houdini_docker_builder/
+commands = pydocstringformatter --write bin/ python/hython_docker_image_builder/
